+++ src/lib/wildcard.ts
